+++ crates/holochain/src/sweettest/mod.rs
@@ -18,6 +18,7 @@ mod sweet_cell;
 mod sweet_conductor;
 mod sweet_conductor_batch;
 mod sweet_conductor_config;
+mod sweet_conductor_config_rendezvous;
 mod sweet_conductor_handle;
 mod sweet_dna;
 mod sweet_network;
@@ -29,6 +30,7 @@ pub use sweet_cell::*;
 pub use sweet_conductor::*;
 pub use sweet_conductor_batch::*;
 pub use sweet_conductor_config::*;
+pub use sweet_conductor_config_rendezvous::*;
 pub use sweet_conductor_handle::*;
 pub use sweet_dna::*;
 pub use sweet_network::*;
